requirements.txt
numpy
pandas
matplotlib
scikit-learn
seaborn

wandb
dotmap
loguru
pyyaml
black
isort
python-dotenv

torch --extra-index-url https://download.pytorch.org/whl/cpu
torchvision --extra-index-url https://download.pytorch.org/whl/cpu
pytorch-lightning
torchmetrics==0.5.1  # https://github.com/Lightning-AI/lightning/issues/4237 removed

pytorch-metric-learning
faiss-cpu

git+https://github.com/IlMioFrizzantinoAmabile/stochman.git#egg=stochman
